Revert "Implement `dependsOn` for invokes in the NodeJS SDK (#16560)"
From SDKs, we call invokes in one of two ways:

* In a "non-output" context (e.g. `getX`), which has a result dependent
  on the language (e.g. a `Promise` in NodeJS) that _does not_ track
  dependencies.
* In an "output" context (e.g. `getXOutput`), which has an `Output` type
  and does track dependencies.

In the non-output case, `dependsOn` really doesn't make sense, since
this style of invoke is inherently ignoring dependency tracking/outputs.
This commit thus reverts 492c57c7dd64c3bf0bc457c33e0be5de3f236bac so
that we can rethink the design before people's programs are subtly
broken in this case.
