+++ changelog/pending/20240712--sdk-nodejs--revert-the-implementation-of-dependson-for-invokes-in-the-nodejs-sdk.yaml
@@ -1,0 +1,4 @@
+changes:
+- type: fix
+  scope: sdk/nodejs
+  description: Revert the implementation of `dependsOn` for invokes in the NodeJS SDK
