+++ sdk/nodejs/invoke.ts
@@ -1,4 +1,4 @@
-// Copyright 2016-2024, Pulumi Corporation.
+// Copyright 2016-2018, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { Input } from "./output";
 import { ProviderResource, Resource } from "./resource";
 
 /*
@@ -52,9 +51,4 @@ export interface InvokeOptions {
      * time.
      */
     async?: boolean;
-
-    /**
-     * An optional set of additional explicit dependencies on other resources.
-     */
-    dependsOn?: Input<Input<Resource>[]> | Input<Resource>;
 }
