--- sdk/nodejs/tests/runtime/langhost/cases/076.invoke_depends_on_non_resource/index.js
@@ -1,6 +1,0 @@
-// This tests that invokes cannot depend on things which are not resources.
-
-let pulumi = require("../../../../../");
-
-const dependsOn = pulumi.output(Promise.resolve([Promise.resolve(1)]));
-pulumi.runtime.invoke("test:index:echo", {}, { dependsOn });
