+++ sdk/nodejs/tsconfig.json
@@ -51,7 +51,6 @@
         "runtime/asyncIterableUtil.ts",
         "runtime/config.ts",
         "runtime/debuggable.ts",
-        "runtime/dependsOn.ts",
         "runtime/invoke.ts",
         "runtime/mocks.ts",
         "runtime/resource.ts",
